# docs/swagger.yaml
definitions:
...
  controllers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  controllers.addressInput:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      latitude:
        type: string
      longitude:
        type: string
      province:
        type: string
      user_id:
        type: integer
      zip:
        type: integer
    type: object
  controllers.cartInput:
    properties:
      note:
        type: string
      qty:
        type: integer
    type: object
...
  models.Address:
    properties:
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      id:
        type: integer
      latitude:
        type: string
      longitude:
        type: string
      province:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      zip:
        type: integer
    type: object
...
  models.Order:
    properties:
      complete_at:
        type: string
      courier:
        type: string
      created_at:
        type: string
      id:
        type: integer
      paid_at:
        type: string
      payment:
        type: string
      resi:
        type: string
      sent_at:
        type: string
      status:
        type: string
      total:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.OrderProduct:
    properties:
      created_at:
        type: string
      id:
        type: integer
      note:
        type: string
      order_id:
        type: integer
      product_id:
        type: integer
      qty:
        type: integer
      total:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
  /address:
    post:
      description: Creating new Address
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create new Address
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.addressInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Address'
      summary: Create new Address
      tags:
      - Buyer
      - Seller
  /address/{id}:
    delete:
      description: Delete address by id
      parameters:
      - description: Address id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete address
      tags:
      - Buyer
      - Seller
    get:
      description: Get Address by Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Address'
      summary: Get Address detail
      tags:
      - Buyer
      - Seller
    put:
      description: Update Address by id
      parameters:
      - description: Address id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update address
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.addressInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Address'
      summary: Update Address
      tags:
      - Buyer
      - Seller
  /cart/{id}/:
    delete:
      description: Delete Product from Order
      parameters:
      - description: OrderProduct id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Delete Product from Order
      tags:
      - Buyer
    put:
      description: Update Product from Order
      parameters:
      - description: OrderProduct id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.cartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderProduct'
      summary: Update Product from Order
      tags:
      - Buyer
...
  /login:
    post:
      description: Search Product, Shop, Category by keyword
      parameters:
      - description: keyword of searching
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Search by keyword
      tags:
      - Public
...
  /products/{id}/add-cart:
    post:
      description: Add Product to Order
      parameters:
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      summary: Add Product to Order
      tags:
      - Buyer
  /register:
    post:
      description: registering user from public access
      parameters:
      - description: the body to register user
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register a user
      tags:
      - Public
...
  /users/address:
    get:
      description: Get a list of Address
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Address'
            type: array
      summary: Get all Address
      tags:
      - Buyer
      - Seller
...
swagger: "2.0"